requirements.txt
groq==0.11.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.38.0
streamlit_mermaid==0.2.0
